feat: 検索ロジックの分離とキーボードショートカットの再定義
